% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwr_basic.R
\name{coef}
\alias{coef}
\alias{coef.gwrm}
\title{Get coefficients of a basic GWR model.}
\usage{
\method{coef}{gwrm}(object, ...)
}
\arguments{
\item{object}{A "gwrm" object.}

\item{\dots}{Additional arguments passing to [coef()].}
}
\description{
Get coefficients of a basic GWR model.
}
