% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwr_basic.R
\name{fitted.gwrm}
\alias{fitted.gwrm}
\title{Get fitted values of a basic GWR model.}
\usage{
\method{fitted}{gwrm}(object, ...)
}
\arguments{
\item{object}{A "gwrm" object.}

\item{\dots}{Additional arguments passing to [fitted()].}
}
\description{
Get fitted values of a basic GWR model.
}
