% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwdr.R, R/gwr_basic.R, R/gwr_multiscale.R
\name{fitted}
\alias{fitted}
\alias{fitted.gwdrm}
\alias{fitted.gwrm}
\alias{fitted.gwrmultiscalem}
\title{Get fitted values of a basic GWR model.}
\usage{
\method{fitted}{gwdrm}(object, ...)

\method{fitted}{gwrm}(object, ...)

\method{fitted}{gwrmultiscalem}(object, ...)
}
\arguments{
\item{object}{A "gwrmultiscalem" object.}

\item{\dots}{Additional arguments passing to [fitted()].}
}
\description{
Get fitted values of a basic GWR model.

Get fitted values of a basic GWR model.

Get fitted values of a basic GWR model.
}
