% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwpca.R
\name{glyph.plot.gwpcam}
\alias{glyph.plot.gwpcam}
\title{Glyph plot for local loadings in PC1 of GWPCA model.}
\usage{
\method{glyph.plot}{gwpcam}(x, y, ..., columns, sign = FALSE)
}
\arguments{
\item{x}{A "gwpcam" object.}

\item{y}{Ignored.}

\item{\dots}{Additional arguments passed to \code{\link[sf:plot]{sf::plot()}}.}

\item{columns}{Column names to plot.
If it is missing or non-character value, all coefficient columns are plotted.}

\item{sign}{whether to plot according to the loadings is beyond 0 or not.}
}
\description{
Glyph plot for local loadings in PC1 of GWPCA model.
}
