% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtdr.R
\name{gtdr}
\alias{gtdr}
\alias{plot.gtdrm}
\alias{step.gtdrm}
\alias{coef.gtdrm}
\alias{fitted.gtdrm}
\alias{residuals.gtdrm}
\title{Geographical and Temporal Density Regression (GTDR)}
\usage{
gtdr(
  formula,
  data,
  config = list(gtdr_config()),
  optim_bw = c("no", "AIC", "CV"),
  optim_bw_threshold = 1e-06,
  optim_bw_step = 0.02,
  optim_bw_max_iter = 1e+06,
  coords = NULL,
  parallel_method = c("no", "omp"),
  parallel_arg = c(0),
  verbose = FALSE
)

\method{plot}{gtdrm}(x, y, ..., columns)

\method{step}{gtdrm}(
  object,
  ...,
  criterion = c("AIC"),
  threshold = 3,
  config = list(gtdr_config()),
  optim_bw = c("no", "AIC", "CV"),
  optim_bw_threshold = 1e-06,
  optim_bw_step = 0.02,
  optim_bw_max_iter = 1e+06
)

\method{coef}{gtdrm}(object, ...)

\method{fitted}{gtdrm}(object, ...)

\method{residuals}{gtdrm}(object, ...)
}
\arguments{
\item{formula}{Regresison model.}

\item{data}{A \code{sf} objects.}

\item{config}{Dimension-specified weighting configuration.
If it is omitted, bandwidth settings will be inherited from \code{object}.
Otherwise, it must be a list of \code{GTDRConfig} objects.
Its length can be 1 or equal to the number of columns of coordinates in \code{data}.
When its length is 1, elements will be duplicated for each column in coordinates.}

\item{optim_bw}{Whether optimize bandwidth after selecting models.
Avaliable values are \code{no}, \code{AIC}, and \code{CV}.
If \code{no} is specified, the bandwidth specified by argument \code{bw}
is used in calibrating selected models.}

\item{optim_bw_threshold}{Threshold of bandwidth optimization.}

\item{optim_bw_step}{Step size in bandwidth optimization.}

\item{optim_bw_max_iter}{Maximum of iteration in bandwidth optimization.}

\item{parallel_method}{Parallel method.}

\item{parallel_arg}{Parallel method argument.}

\item{verbose}{Whether to print additional information.
A higher value leads to more information.}

\item{x}{A "gtdrm" object.}

\item{y}{Ignored.}

\item{\dots}{Additional arguments passing to \code{\link[=residuals]{residuals()}}.}

\item{columns}{Column names to plot.
If it is missing or non-character value, all coefficient columns are plottd.}

\item{object}{A "gtdrm" object.}

\item{criterion}{The model-selection method.
Currently there is only AIC available.}

\item{threshold}{The threshold of criterion changes. Default to 3.0.}
}
\value{
A \code{gtdrm} object.
}
\description{
Geographical and Temporal Density Regression (GTDR)
}
\section{Functions}{
\itemize{
\item \code{plot(gtdrm)}: Plot the result of GTDR model.

\item \code{step(gtdrm)}: Model selection for GTDR model

\item \code{coef(gtdrm)}: Get coefficients of a GTDR model.

\item \code{fitted(gtdrm)}: Get fitted values of a GTDR model.

\item \code{residuals(gtdrm)}: Get residuals of a GTDR model.

}}
\examples{
data(LondonHP)
gtdr(PURCHASE ~ FLOORSZ + UNEMPLOY, LondonHP)

### Specific Bandwidth
gtdr(PURCHASE ~ FLOORSZ + UNEMPLOY, LondonHP, list(
    gtdr_config(0.2, TRUE, "gaussian"),
    gtdr_config(0.2, TRUE, "gaussian")
))

### Optim Bandwidth
m <- gtdr(PURCHASE ~ FLOORSZ + UNEMPLOY, LondonHP, list(
    gtdr_config(0.618, TRUE, "gaussian")
), optim_bw = "AIC")
m

plot(m)

step(m, threshold = 100.0)

coef(m)

fitted(m)

residuals(m)

}
