% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-GTDRConfig.R
\docType{class}
\name{GTDRConfig-class}
\alias{GTDRConfig-class}
\alias{GTDRConfig}
\title{An S4 class to set GTDR configurations}
\description{
An S4 class to set GTDR configurations
}
\section{Slots}{

\describe{
\item{\code{bw}}{Bandwidth value.}

\item{\code{adaptive}}{Whether the bandwidth value is adaptive or not.}

\item{\code{kernel}}{Kernel function used.}
}}

