% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtwr.R
\name{gtwr}
\alias{gtwr}
\title{Geographically and Temporally Weighted Regression}
\usage{
gtwr(
  formula,
  data,
  timestamps,
  bw = NA,
  lambda = NA,
  adaptive = FALSE,
  kernel = c("gaussian", "exp", "bisquare", "tricube", "boxcar"),
  longlat = FALSE,
  p = 2,
  theta = 0,
  optim_bw_range = c(0, Inf),
  hatmatrix = TRUE,
  time_format = NA,
  parallel_method = c("no", "omp"),
  parallel_arg = c(0),
  verbose = FALSE
)
}
\arguments{
\item{formula}{Formula for regression.}

\item{data}{A \code{sf} objects.}

\item{timestamps}{A vector timestamps for all samples.
Either a numerical vector, or a character vector to be parsed according to \code{time_format}.}

\item{bw}{Either a value to set the size of bandwidth,
or one of the following characters to set the criterion for
bandwidth auto-optimization process.
\itemize{
\item \code{AIC}
\item \code{CV}
Note that if \code{NA} or other non-numeric value is setted,
this parameter will be reset to \code{Inf}.
}}

\item{lambda}{Either a value between 0 and 1 as the weight of temporal distance,
or a \code{NA} value to enable auto-selection.}

\item{adaptive}{Whether the bandwidth value is adaptive or not.}

\item{kernel}{Kernel function used.}

\item{longlat}{Whether the coordinates}

\item{p}{Power of the Minkowski distance,
default to 2, i.e., Euclidean distance.}

\item{theta}{Angle in radian to roate the coordinate system, default to 0.}

\item{optim_bw_range}{Bounds on bandwidth optimization, a vector of two numeric elements.
Set to \code{NA_real_} to enable default values selected by the algorithm.}

\item{hatmatrix}{If TRUE, great circle will be caculated.}

\item{time_format}{The format used to parse timestamps if they are characters.}

\item{parallel_method}{Parallel method.}

\item{parallel_arg}{Parallel method argument.}

\item{verbose}{Whether to print additional information.}
}
\value{
A \code{gtwrm} object.
}
\description{
Geographically and Temporally Weighted Regression
}
\details{
\subsection{Parallelization}{

Supported method(s):
\itemize{
\item Multithreading (\code{omp})
}

See the vignettes about parallelization to learn more about this topic.
}
}
\examples{
data(LondonHP)
LondonHP$time <- as.integer(round(runif(nrow(LondonHP), 1, 365)))

# Basic usage
m <- gtwr(PURCHASE~FLOORSZ+UNEMPLOY, LondonHP, "time", 64, 0.05, TRUE)
m
coef(m)
fitted(m)
residual(m)

# Bandwidth Optimization
gtwr(PURCHASE~FLOORSZ+UNEMPLOY, LondonHP, LondonHP$time, NA, 0.05, TRUE)

# Lambda Optimization
gtwr(PURCHASE~FLOORSZ+UNEMPLOY, LondonHP, LondonHP$time, 64, NA, TRUE)

# Bandwidth and Lambda optimization
gtwr(PURCHASE~FLOORSZ+UNEMPLOY, LondonHP, LondonHP$time, NA, NA, TRUE)

}
