% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwss.R
\name{gw.average}
\alias{gw.average}
\title{Geographically Weighted Average}
\usage{
gw.average(
  data,
  vars,
  bw,
  adaptive = FALSE,
  quantile = FALSE,
  kernel = c("gaussian", "exp", "bisquare", "tricube", "boxcar"),
  longlat = FALSE,
  p = 2,
  theta = 0,
  parallel_method = c("no", "omp"),
  parallel_arg = c(0)
)
}
\arguments{
\item{data}{A \code{sf} objects.}

\item{vars}{a vector of variable names to be summarized.}

\item{bw}{Bandwidth value, need to be specified.}

\item{adaptive}{Whether the bandwidth value is adaptive or not.}

\item{quantile}{Whether to calculate local quantiles.}

\item{kernel}{Kernel function used.}

\item{longlat}{Whether the coordinates}

\item{p}{Power of the Minkowski distance,
default to 2, i.e., Euclidean distance.}

\item{theta}{Angle in radian to roate the coordinate system, default to 0.}

\item{parallel_method}{Parallel method.}

\item{parallel_arg}{Parallel method argument.}
}
\value{
A \code{gwssm} object.
}
\description{
Geographically Weighted Average
}
\examples{
data(LondonHP)
m <- gw.average(LondonHP, c("PURCHASE","FLOORSZ","UNEMPLOY"), 64, TRUE)
gw.average(LondonHP, "PURCHASE,FLOORSZ,UNEMPLOY", 64, TRUE)

}
