% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwss.R
\name{gw.correlation}
\alias{gw.correlation}
\title{Geographically Weighted Correlation}
\usage{
gw.correlation(
  data,
  vars1,
  vars2,
  bws = NULL,
  adaptive = FALSE,
  kernel = c("gaussian", "exp", "bisquare", "tricube", "boxcar"),
  approach = c("CV", "AIC"),
  longlat = FALSE,
  p = 2,
  theta = 0,
  parallel_method = c("no", "omp"),
  parallel_arg = c(0)
)
}
\arguments{
\item{data}{A \code{sf} objects.}

\item{vars1}{a vector of variable names to be calculated.}

\item{vars2}{a vector of variable names as the responsed.}

\item{bws}{Bandwidth value list, auto-select if not provided.}

\item{adaptive}{Whether the bandwidth value is adaptive or not.}

\item{kernel}{Kernel function used.}

\item{approach}{Bandwidth selection.}

\item{longlat}{Whether the coordinates}

\item{p}{Power of the Minkowski distance,
default to 2, i.e., Euclidean distance.}

\item{theta}{Angle in radian to roate the coordinate system, default to 0.}

\item{parallel_method}{Parallel method.}

\item{parallel_arg}{Parallel method argument.}
}
\value{
A \code{gwssm} object.
}
\description{
Geographically Weighted Correlation
}
\examples{
data(LondonHP)
m <- gw.correlation(LondonHP, c("PURCHASE","PROF"),c("FLOORSZ","UNEMPLOY"),kernel="bisquare", adaptive=TRUE)
gw.correlation(LondonHP, c("PURCHASE"),c("FLOORSZ","UNEMPLOY"),c(50,50), adaptive=TRUE)

}
