% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gw_correlation.R
\name{gw_correlation}
\alias{gw_correlation}
\title{Geographically Weighted Correlation}
\usage{
gw_correlation(
  formula,
  data,
  config = list(gwcorr_config()),
  parallel_method = c("no", "omp"),
  parallel_arg = c(0)
)
}
\arguments{
\item{data}{A \code{sf} objects.}

\item{parallel_method}{Parallel method.}

\item{parallel_arg}{Parallel method argument.}

\item{vars1}{a vector of variable names to be calculated.}

\item{vars2}{a vector of variable names as the responsed.}

\item{bws}{Bandwidth value list, auto-select if not provided.}

\item{adaptive}{Whether the bandwidth value is adaptive or not.}

\item{kernel}{Kernel function used.}

\item{approach}{Bandwidth selection.}

\item{longlat}{Whether the coordinates}

\item{p}{Power of the Minkowski distance,
default to 2, i.e., Euclidean distance.}

\item{theta}{Angle in radian to roate the coordinate system, default to 0.}
}
\value{
A \code{gwcorrm} object.
}
\description{
Geographically Weighted Correlation
}
\examples{
data(LondonHP)
gw_correlation(
 formula = PURCHASE ~ FLOORSZ + UNEMPLOY + PROF,
 data = LondonHP
)
# Specify more variable pairs
gw_correlation(
 formula = PURCHASE + FLOORSZ ~ UNEMPLOY + PROF,
 data = LondonHP
)

# Specify more configurations for all variables
gw_correlation(
 formula = PURCHASE + FLOORSZ ~ UNEMPLOY + PROF,
 data = LondonHP,
 config = list(gwcorr_config(adaptive = TRUE, kernel = "bisquare"))
)
m

# Specify more configurations for each variables
gw_correlation(
 formula = PURCHASE + FLOORSZ ~ UNEMPLOY + PROF,
 data = LondonHP,
 config = list(
     gwcorr_config(adaptive = TRUE, kernel = "gaussian"),
     gwcorr_config(adaptive = TRUE, kernel = "gaussian"),
     gwcorr_config(adaptive = TRUE, kernel = "bisquare"),
     gwcorr_config(adaptive = TRUE, kernel = "bisquare")
 ))

# Specify configurations by variable names, names should be paired with '_'
gw_correlation(
 formula = PURCHASE + FLOORSZ ~ UNEMPLOY + PROF,
 data = LondonHP,
 config = list(
     PURCHASE_UNEMPLOY = gwcorr_config(bw = 20, adaptive = TRUE, kernel = "bisquare"),
     .default = gwcorr_config(adaptive = TRUE, kernel = "bisquare")
 ))

}
