% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-GWCorrConfig.R
\name{gwcorr_config}
\alias{gwcorr_config}
\title{Create an instance of GWCorrConfig}
\usage{
gwcorr_config(
  bw = NA_real_,
  adaptive = FALSE,
  kernel = c("gaussian", "exp", "bisquare", "tricube", "boxcar"),
  longlat = FALSE,
  p = 2,
  theta = 0,
  optim_bw = c("CV", "AIC", "no")
)
}
\arguments{
\item{bw}{Bandwidth value.}

\item{adaptive}{Whether the bandwidth value is adaptive or not.}

\item{kernel}{Kernel function used.}

\item{longlat}{Whether the coordinates.}

\item{p}{Power of the Minkowski distance,
default to 2, i.e., Euclidean distance.}

\item{theta}{Angle in radian to roate the coordinate system, default to 0.}

\item{optim_bw}{Whether optimize bandwidth after selecting models.
Avaliable values are \code{no}, \code{AIC}, and \code{CV}.
If \code{no} is specified, the bandwidth specified by argument \code{bw}
is used in calibrating selected models.}
}
\description{
Create an instance of GWCorrConfig
}
\examples{
gwcorr_config(36, TRUE, "bisquare", optim_bw = "AIC")

}
