% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-GWCorrConfig.R
\docType{class}
\name{GWCorrConfig-class}
\alias{GWCorrConfig-class}
\alias{GWCorrConfig}
\title{An S4 class to set GW Correlation configurations}
\description{
An S4 class to set GW Correlation configurations
}
\section{Slots}{

\describe{
\item{\code{bw}}{Bandwidth value.}

\item{\code{adaptive}}{Whether the bandwidth value is adaptive or not.}

\item{\code{kernel}}{Kernel function used.}

\item{\code{longlat}}{Whether the coordinates.}

\item{\code{p}}{Power of the Minkowski distance,
default to 2, i.e., Euclidean distance.}

\item{\code{theta}}{Angle in radian to roate the coordinate system, default to 0.}

\item{\code{optim_bw}}{Whether optimize bandwidth after selecting models.
Avaliable values are \code{no}, \code{AIC}, and \code{CV}.
If \code{no} is specified, the bandwidth specified by argument \code{bw}
is used in calibrating selected models.}
}}

