% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwcorrelation.R
\name{gwcorrelation}
\alias{gwcorrelation}
\title{Geographically Weighted Correlation}
\usage{
gwcorrelation(
  formula,
  data,
  config = list(gwcorr_config()),
  parallel_method = c("no", "omp"),
  parallel_arg = c(0),
  verbose = FALSE
)
}
\arguments{
\item{formula}{A formula, seperated by '~' and joined by '+'}

\item{data}{A \code{sf} objects.}

\item{config}{Parameter-specified weighting configuration.
It must be a list of \code{GWCorr} objects.
Please find more details in the details section.}

\item{parallel_method}{Parallel method.}

\item{parallel_arg}{Parallel method argument.}

\item{verbose}{Whether to print additional information.}
}
\value{
A \code{gwcorrm} object.
}
\description{
Geographically Weighted Correlation
}
\examples{
data(LondonHP)
# Basic use
m <- gwcorrelation(formula = PURCHASE ~ FLOORSZ + UNEMPLOY, data = LondonHP)
# Specify more variable pairs
gwcorrelation(
 formula = PURCHASE + FLOORSZ ~ UNEMPLOY + PROF,
 data = LondonHP
)
# Specify using default for all variables
gwcorrelation(
  formula = PURCHASE + FLOORSZ ~ UNEMPLOY + PROF,
  data = LondonHP,
  config = list(.default = gwcorr_config(adaptive = TRUE, kernel = "gaussian", optim_bw="CV"))
)
# or
gwcorrelation(
 formula = PURCHASE + FLOORSZ ~ UNEMPLOY + PROF,
 data = LondonHP,
 config = list(gwcorr_config(adaptive = FALSE, kernel = "bisquare"))
)

# Specify more configurations for each variables
# Make sure the numbers of variables are correct
gwcorrelation(
 formula = PURCHASE + FLOORSZ ~ UNEMPLOY + PROF,
 data = LondonHP,
 config = list(
     gwcorr_config(adaptive = TRUE, kernel = "gaussian"),
     gwcorr_config(adaptive = TRUE, kernel = "gaussian"),
     gwcorr_config(adaptive = TRUE, kernel = "bisquare"),
     gwcorr_config(adaptive = TRUE, kernel = "bisquare")
 ))
# or names should be paired with '_'
gwcorrelation(
 formula = PURCHASE + FLOORSZ ~ UNEMPLOY + PROF,
 data = LondonHP,
 config = list(
     PURCHASE_UNEMPLOY = gwcorr_config(kernel="bisquare", adaptive = TRUE),
     PURCHASE_PROF = gwcorr_config(kernel="bisquare", adaptive = FALSE),
     FLOORSZ_UNEMPLOY = gwcorr_config(kernel="gaussian", adaptive = TRUE),
     FLOORSZ_PROF = gwcorr_config(kernel="gaussian", adaptive = FALSE)
 ))

# Specify configurations by variable names and default
gwcorrelation(
 formula = PURCHASE + FLOORSZ ~ UNEMPLOY + PROF,
 data = LondonHP,
 config = list(
     PURCHASE_UNEMPLOY = gwcorr_config(bw = 20, adaptive = TRUE, kernel = "bisquare"),
     .default = gwcorr_config(adaptive = TRUE, kernel = "bisquare")
 ))

}
