% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwda.R
\name{gwda}
\alias{gwda}
\title{Geographically Weighted Summary Statistics}
\usage{
gwda(
  formula,
  data,
  bw,
  adaptive = FALSE,
  kernel = c("gaussian", "exp", "bisquare", "tricube", "boxcar"),
  longlat = FALSE,
  p = 2,
  theta = 0,
  method = c("wqda", "wlda"),
  parallel_method = c("no", "omp"),
  parallel_arg = c(0)
)
}
\arguments{
\item{formula}{Regresison model.}

\item{data}{A \code{sf} objects.}

\item{bw}{Bandwidth value}

\item{adaptive}{Whether the bandwidth value is adaptive or not.}

\item{kernel}{Kernel function used.}

\item{longlat}{Whether the coordinates}

\item{p}{Power of the Minkowski distance,
default to 2, i.e., Euclidean distance.}

\item{theta}{Angle in radian to roate the coordinate system, default to 0.}

\item{parallel_method}{Parallel method.}

\item{parallel_arg}{Parallel method argument.}

\item{quantile}{Whether to calculate local quantiles.}
}
\value{
A \code{gwdam} object.
}
\description{
Geographically Weighted Summary Statistics
}
