% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwdr.R
\name{gwdr}
\alias{gwdr}
\title{Geographically Weighted Density Regression}
\usage{
gwdr(
  formula,
  data,
  config = list(gwdr_config()),
  optim_bw = c("no", "AIC", "CV"),
  optim_bw_threshold = 1e-06,
  optim_bw_step = 0.02,
  optim_bw_max_iter = 1e+06,
  parallel_method = c("no", "omp"),
  parallel_arg = c(0)
)
}
\description{
Geographically Weighted Density Regression
}
\examples{
data(LondonHP)
gwdr(PURCHASE~FLOORSZ+UNEMPLOY, LondonHP, list(
    gwdr_config(0.618, T)
))

### Optim Bandwidth
data(LondonHP)
gwdr(PURCHASE~FLOORSZ+UNEMPLOY, LondonHP, list(
    gwdr_config(0.618, T)
), optim_bw = "AIC")

}
