% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-GWDRConfig.R
\docType{class}
\name{GWDRConfig-class}
\alias{GWDRConfig-class}
\alias{GWDRConfig}
\alias{gwdr_config}
\title{An S4 class to set GWDR configurations}
\usage{
gwdr_config(
  bw = 0.618,
  adaptive = TRUE,
  kernel = c("gaussian", "exp", "bisquare", "tricube", "boxcar")
)
}
\description{
An S4 class to set GWDR configurations

Create an instance of GWDRConfig
}
\section{Functions}{
\itemize{
\item \code{gwdr_config()}: 

}}
\section{Slots}{

\describe{
\item{\code{bw}}{Bandwidth Value}
}}

\examples{
gwdr_config(36, TRUE, "bisquare", optim_bw = "AIC")

}
