% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWmodel3-package.R
\docType{package}
\name{GWmodel3-package}
\alias{GWmodel3}
\alias{GWmodel3-package}
\title{GWmodel3: Modelling Spatial Data with Geographically Weighted Models}
\description{
Techniques from a particular branch of spatial statistics, termed geographically-weighted (GW) models. GW models suit situations when data are not described well by some global model, but where there are spatial regions where a suitably localised calibration provides a better description. 'GWmodel3' includes: GW summary statistics (Brunsdon et al., 2002)\doi{10.1016/s0198-9715(01)00009-6}, GW principal components analysis (Harris et al., 2011)\doi{10.1080/13658816.2011.554838}, GW discriminant analysis (Brunsdon et al., 2007)\doi{10.1111/j.1538-4632.2007.00709.x} and various forms of GW regression (Brunsdon et al., 1996)\doi{10.1111/j.1538-4632.1996.tb00936.x}; some of which are provided in basic and robust (outlier resistant) forms.
}
\details{
In GWmodel, we introduce techniques from a particular branch of spatial statistics,
termed geographically-weighted (GW) models. GW models suit situations when data are
not described well by some global model, but where there are spatial regions where
a suitably localised calibration provides a better description. GWmodel includes functions
to calibrate: GW summary statistics, GW principal components analysis, GW discriminant analysis
and various forms of GW regression; some of which are provided in basic and robust (outlier resistant) forms.
In particular, the high-performence computing technologies, including multi-thread and CUDA techniques are started
to be adopted for efficient calibrations.
}
\note{
Acknowledgements: We gratefully acknowledge support from National Natural Science Foundation of China (42071368);
Science Foundation Ireland under the National Development Plan through the award of a Strategic Research Centre grant 07-SRC-I1168.

For latest tutorials on using GWmodel please go to: \url{https://rpubs.com/gwmodel}
}
\references{
Gollini I, Lu B, Charlton M, Brunsdon C, Harris P (2015) GWmodel: an R Package
for exploring Spatial Heterogeneity using Geographically Weighted Models. Journal
of Statistical Software, 63(17):1-50, \doi{10.18637/jss.v063.i17}

Lu B, Harris P, Charlton M, Brunsdon C (2014) The GWmodel R Package: further
topics for exploring Spatial Heterogeneity using Geographically Weighted Models.
Geo-spatial Information Science 17(2): 85-101,
\doi{10.1080/10095020.2014.917453}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/GWmodel-Lab/GWmodel3}
  \item Report bugs at \url{https://github.com/GWmodel-Lab/GWmodel3/issues}
}

}
\author{
\strong{Maintainer}: Binbin Lu \email{binbinlu@whu.edu.cn} (\href{https://orcid.org/0000-0001-7847-7560}{ORCID})

Authors:
\itemize{
  \item Yigong Hu \email{yigong.hu@bristol.ac.uk} (\href{https://orcid.org/0000-0002-9553-6275}{ORCID})
}

Other contributors:
\itemize{
  \item Haotian Zhang [contributor]
  \item Guangyu Ou [contributor]
  \item Paul Harris [contributor]
  \item Martin Charlton [contributor]
  \item Chris Brunsdon [contributor]
  \item Tomoki Nakaya [contributor]
  \item Daisuke Murakami [contributor]
  \item Isabella Gollini [contributor]
  \item Fiona H Evans [contributor]
}

}
\keyword{internal}
