% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwpca.R
\name{gwpca}
\alias{gwpca}
\title{Calibrate a GWPCA model}
\usage{
gwpca(
  formula,
  data,
  bw = NA,
  adaptive = FALSE,
  kernel = c("gaussian", "exp", "bisquare", "tricube", "boxcar"),
  longlat = FALSE,
  p = 2,
  theta = 0,
  components = 2,
  scale = FALSE
)
}
\arguments{
\item{formula}{Regresison model.}

\item{data}{A \code{sf} objects.}

\item{bw}{Should provide a value to set the size of bandwidth,}

\item{adaptive}{Whether the bandwidth value is adaptive or not.}

\item{kernel}{Kernel function used.}

\item{longlat}{Whether the coordinates}

\item{p}{Power of the Minkowski distance,
default to 2, i.e., Euclidean distance.}

\item{theta}{Angle in radian to roate the coordinate system, default to 0.}

\item{components}{How many components want to keep, default to 2.}

\item{scale}{Whether to standardised data.}
}
\value{
A \code{gwpcam} object.
}
\description{
Calibrate a GWPCA model
}
\examples{
data(LondonHP)

# Basic usage
gwpca(~FLOORSZ + UNEMPLOY + PROF, LondonHP, 50, TRUE, components = 2)

}
\seealso{
\code{browseVignettes("")}
}
