% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwr_basic.R
\name{gwr_basic}
\alias{gwr_basic}
\title{Calibrate a basic GWR model}
\usage{
gwr_basic(
  formula,
  data,
  bw,
  adaptive = FALSE,
  kernel = c("gaussian", "exp", "bisquare", "tricube", "boxcar"),
  longlat = FALSE,
  p = 2,
  theta = 0,
  hatmatrix = TRUE,
  parallel_method = c("none", "omp", "cuda", "cluster"),
  parallel_arg = c(0)
)
}
\arguments{
\item{formula}{Regresison model.}

\item{data}{A `sf` objects.}

\item{bw}{Bandwidth value. Depends on the value of `adaptive`.}

\item{adaptive}{Whether the bandwidth value is adaptive or not.}

\item{kernel}{Kernel function used.}

\item{longlat}{Whether the coordinates}

\item{p}{Power of the Minkowski distance, default is 2, i.e. the Euclidean distance.}

\item{theta}{Angle in radian to roate the coordinate system, default to 0.}

\item{hatmatrix}{If TRUE, great circle will be caculated}

\item{parallel_method}{Parallel method.}

\item{parallel_arg}{Parallel method argument.}
}
\value{
A `gwrm` object.
}
\description{
Calibrate a basic GWR model
}
\examples{
data(LondonHP)
gwr_basic(PURCHASE~FLOORSZ+UNEMPLOY, LondonHP, 64, TRUE)

}
