% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwr_generalized.R
\name{gwr_generalized}
\alias{gwr_generalized}
\alias{plot.ggwrm}
\alias{coef.ggwrm}
\title{Calibrate a basic GGWR model}
\usage{
gwr_generalized(
  formula,
  data,
  family = c("poisson", "binomial"),
  bw = NA,
  adaptive = FALSE,
  kernel = c("gaussian", "exp", "bisquare", "tricube", "boxcar"),
  longlat = FALSE,
  p = 2,
  theta = 0,
  hatmatrix = TRUE,
  parallel_method = c("no", "omp"),
  parallel_arg = c(0)
)

\method{plot}{ggwrm}(x, y, ..., columns)

\method{coef}{ggwrm}(object, ...)
}
\arguments{
\item{formula}{Regresison model.}

\item{data}{A \code{sf} objects.}

\item{bw}{Either a value to set the size of bandwidth,
or one of the following characters to set the criterion for
bandwidth auto-optimization process.
\itemize{
\item \code{AIC}
\item \code{CV}
Note that if \code{NA} or other non-numeric value is setted,
this parameter will be reset to \code{Inf}.
}}

\item{adaptive}{Whether the bandwidth value is adaptive or not.}

\item{kernel}{Kernel function used.}

\item{longlat}{Whether the coordinates}

\item{p}{Power of the Minkowski distance,
default to 2, i.e., Euclidean distance.}

\item{theta}{Angle in radian to roate the coordinate system, default to 0.}

\item{hatmatrix}{If TRUE, great circle will be caculated.}

\item{parallel_method}{Parallel method.}

\item{parallel_arg}{Parallel method argument.}

\item{x}{A "ggwrm" object.}

\item{y}{Ignored.}

\item{\dots}{Additional arguments passing to \code{\link[=coef]{coef()}}.}

\item{columns}{Column names to plot.
If it is missing or non-character value, all coefficient columns are plottd.}

\item{object}{A "ggwrm" object.}

\item{optim_bw_range}{Bounds on bandwidth optimization, a vector of two numeric elements.
Set to \code{NA_real_} to enable default values selected by the algorithm.}

\item{verbose}{Whether to print additional information.}
}
\value{
A \code{ggwrm} object.
}
\description{
Calibrate a basic GGWR model
}
\details{
\subsection{Parallelization}{

One parallel methods are provided to speed up basic GGWR algorithm:
\itemize{
\item Multithreading (\code{omp})
}

See the vignettes about parallelization to learn more about this topic.
}
}
\section{Functions}{
\itemize{
\item \code{plot(ggwrm)}: Plot the result of basic GGWR model.

\item \code{coef(ggwrm)}: Get coefficients of a Generalized GWR model.

}}
\examples{
data(LondonHP)

# Basic usage
gwr_generalized(PURCHASE~FLOORSZ+UNEMPLOY, LondonHP, "poisson", 64, TRUE)

# Bandwidth Optimization
m <- gwr_generalized(PURCHASE ~ FLOORSZ + UNEMPLOY + PROF, LondonHP, 'AIC', TRUE)
m

plot(m)

coef(m)

}
\seealso{
\code{browseVignettes("")}
}
