% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwr_localcollinearity.R
\name{gwr_lcr}
\alias{gwr_lcr}
\alias{plot.gwlcrm}
\alias{coef.gwlcrm}
\alias{fitted.gwlcrm}
\alias{residuals.gwlcrm}
\title{Calibrate a basic GWR model}
\usage{
gwr_lcr(
  formula,
  data,
  bw = NA,
  adaptive = FALSE,
  kernel = c("gaussian", "exp", "bisquare", "tricube", "boxcar"),
  longlat = FALSE,
  p = 2,
  theta = 0,
  lambda = 0,
  lambda_adjust = FALSE,
  cn_thresh = 30,
  cv = TRUE,
  hatmatrix = TRUE,
  parallel_method = c("no", "omp"),
  parallel_arg = c(0),
  verbose = FALSE
)

\method{plot}{gwlcrm}(x, y, ..., columns)

\method{coef}{gwlcrm}(object, ...)

\method{fitted}{gwlcrm}(object, ...)

\method{residuals}{gwlcrm}(object, ...)
}
\arguments{
\item{formula}{Regresison model.}

\item{data}{A \code{sf} objects.}

\item{bw}{Either a value to set the size of bandwidth,
or one of the following characters to set the criterion for
bandwidth auto-optimization process.
\itemize{
\item \code{AIC}
\item \code{CV}
Note that if \code{NA} or other non-numeric value is setted,
this parameter will be reset to \code{Inf}.
}}

\item{adaptive}{Whether the bandwidth value is adaptive or not.}

\item{kernel}{Kernel function used.}

\item{longlat}{Whether the coordinates}

\item{p}{Power of the Minkowski distance,
default to 2, i.e., Euclidean distance.}

\item{theta}{Angle in radian to roate the coordinate system, default to 0.}

\item{hatmatrix}{If TRUE, great circle will be caculated.}

\item{parallel_method}{Parallel method.}

\item{parallel_arg}{Parallel method argument.}

\item{verbose}{Whether to print additional information.}

\item{x}{A "gwlcrm" object.}

\item{y}{Ignored.}

\item{\dots}{Additional arguments passing to \code{\link[=residuals]{residuals()}}.}

\item{columns}{Column names to plot.
If it is missing or non-character value, all coefficient columns are plottd.}

\item{object}{A "gwlcrm" object.}

\item{optim_bw_range}{Bounds on bandwidth optimization, a vector of two numeric elements.
Set to \code{NA_real_} to enable default values selected by the algorithm.}
}
\value{
A \code{gwlcrm} object.
}
\description{
Calibrate a basic GWR model
}
\details{
\subsection{Parallelization}{

Two parallel methods are provided to speed up basic GWR algorithm:
\itemize{
\item Multithreading (\code{omp})
\item NVIDIA GPU Computing (\code{cuda})
}

See the vignettes about parallelization to learn more about this topic.
}
}
\section{Functions}{
\itemize{
\item \code{plot(gwlcrm)}: Plot the result of basic GWR model.

\item \code{coef(gwlcrm)}: Get coefficients of a basic GWR model.

\item \code{fitted(gwlcrm)}: Get fitted values of a basic GWR model.

\item \code{residuals(gwlcrm)}: Get residuals of a basic GWR model.

}}
\examples{
data(LondonHP)

# Basic usage
gwr_lcr(PURCHASE ~ FLOORSZ + UNEMPLOY, LondonHP, 64, TRUE)

# Bandwidth Optimization
m <- gwr_lcr(PURCHASE ~ FLOORSZ + UNEMPLOY + PROF, LondonHP, 'AIC', TRUE)
m

plot(m)

coef(m)

fitted(m)

residuals(m)

}
\seealso{
\code{browseVignettes("")}
}
