% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwr_robust.R
\name{gwr_robust}
\alias{gwr_robust}
\alias{step.rgwrm}
\title{Calibrate a Robust GWR model}
\usage{
gwr_robust(
  formula,
  data,
  bw = NA,
  adaptive = FALSE,
  kernel = c("gaussian", "exp", "bisquare", "tricube", "boxcar"),
  longlat = FALSE,
  p = 2,
  theta = 0,
  optim_bw_range = c(0, Inf),
  hatmatrix = TRUE,
  filter = FALSE,
  parallel_method = c("no", "omp", "cuda"),
  parallel_arg = c(0),
  verbose = FALSE
)

\method{step}{rgwrm}(
  object,
  ...,
  criterion = c("AIC"),
  threshold = 3,
  bw = NA,
  optim_bw = c("no", "AIC", "CV")
)
}
\arguments{
\item{formula}{Regresison model.}

\item{data}{A \code{sf} objects.}

\item{bw}{The bandwidth used in selecting models.
If \code{NA} or missing, the bandwidth value will be derived
from the \code{gwrm} object.}

\item{adaptive}{Whether the bandwidth value is adaptive or not.}

\item{kernel}{Kernel function used.}

\item{longlat}{Whether the coordinates}

\item{p}{Power of the Minkowski distance,
default to 2, i.e., Euclidean distance.}

\item{theta}{Angle in radian to roate the coordinate system, default to 0.}

\item{optim_bw_range}{Bounds on bandwidth optimization, a vector of two numeric elements.
Set to \code{NA_real_} to enable default values selected by the algorithm.}

\item{hatmatrix}{If TRUE, great circle will be caculated.}

\item{filter}{Whether to use filtered algorithm.}

\item{parallel_method}{Parallel method.}

\item{parallel_arg}{Parallel method argument.}

\item{verbose}{Whether to print additional information.}

\item{object}{A "\code{gwrm}" class object.}

\item{\dots}{Other parameters.}

\item{criterion}{The model-selection method.
Currently there is only AIC available.}

\item{threshold}{The threshold of criterion changes. Default to 3.0.}

\item{optim_bw}{Whether optimize bandwidth after selecting models.
Avaliable values are \code{no}, \code{AIC}, and \code{CV}.
If \code{no} is specified, the bandwidth specified by argument \code{bw}
is used in calibrating selected models.}
}
\value{
A \code{rgwrm} object.

A \code{rgwrm} object.
}
\description{
Calibrate a Robust GWR model
}
\details{
\subsection{Parallelization}{

Multithreading (\code{omp}) are provided to speed up robust GWR algorithm:

See the vignettes about parallelization to learn more about this topic.
}
}
\section{Functions}{
\itemize{
\item \code{step(rgwrm)}: Model selection for Robust GWR model

}}
\examples{
data(LondonHP)

# Basic usage
gwr_robust(PURCHASE ~ FLOORSZ + UNEMPLOY, LondonHP, 64, TRUE)

# Bandwidth Optimization
m <- gwr_robust(PURCHASE ~ FLOORSZ + UNEMPLOY + PROF, LondonHP, 'AIC', TRUE)
m

step(m, threshold = 100.0, bw = Inf, optim_bw = "AIC")

}
\seealso{
\code{browseVignettes("")}
}
