% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwr_scalable.R
\name{gwr_scalable}
\alias{gwr_scalable}
\title{Geographically Weighted Summary Statistics}
\usage{
gwr_scalable(
  formula,
  data,
  bw,
  adaptive = FALSE,
  kernel = c("gaussian", "exp", "bisquare", "tricube", "boxcar"),
  longlat = FALSE,
  p = 2,
  theta = 0,
  polynomial = 3,
  hatmatrix = TRUE
)
}
\arguments{
\item{formula}{Regresison model.}

\item{data}{A \code{sf} objects.}

\item{bw}{Bandwidth value}

\item{adaptive}{Whether the bandwidth value is adaptive or not.}

\item{kernel}{Kernel function used.}

\item{longlat}{Whether the coordinates}

\item{p}{Power of the Minkowski distance,
default to 2, i.e., Euclidean distance.}

\item{theta}{Angle in radian to roate the coordinate system, default to 0.}

\item{polynomial}{Polynomial}

\item{hatmatrix}{Hatmatrix}
}
\value{
A \code{gwscalem} object.
}
\description{
Geographically Weighted Summary Statistics
}
