% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_table.R
\name{matrix2char}
\alias{matrix2char}
\title{Convert a numeric matrix to character matrix according to a format string.}
\usage{
matrix2char(m, fmt = "\%.3f")
}
\arguments{
\item{m}{A numeric matrix.}

\item{fmt}{Format string. Passing to [base::sprintf()].}
}
\description{
Convert a numeric matrix to character matrix according to a format string.
}
\seealso{
[base::sprintf()], [print.hgwrm()], [print.summary.hgwrm()].
}
