% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwr_multiscale.R
\docType{class}
\name{MGWRConfig-class}
\alias{MGWRConfig-class}
\alias{MGWRConfig}
\title{An S4 class to set Multiscale GWR configurations}
\description{
An S4 class to set Multiscale GWR configurations
}
\section{Slots}{

\describe{
\item{\code{bw}}{Bandwidth Value}
}}

