% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwr_basic.R
\name{model_sel.gwrm}
\alias{model_sel.gwrm}
\title{Model selection for basic GWR model}
\usage{
\method{model_sel}{gwrm}(
  gwrm,
  criterion = c("AIC"),
  threshold = 3,
  bw = NA,
  optim_bw = c("no", "AIC", "CV"),
  ...
)
}
\arguments{
\item{gwrm}{A "`gwrm`" class object.}

\item{criterion}{The model-selection method.
Currently there is only AIC available.}

\item{threshold}{The threshold of criterion changes. Default to 3.0.}

\item{bw}{The bandwidth used in selecting models.
If `NA` or missing, the bandwidth value will be derived
from the `gwrm` object.}

\item{optim_bw}{Whether optimize bandwidth after selecting models.
Avaliable values are `no`, `AIC`, and `CV`.
If `no` is specified, the bandwidth specified by argument `bw`
is used in calibrating selected models.}

\item{\dots}{Other parameters.}
}
\description{
Model selection for basic GWR model
}
