% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwr_basic.R, R/model_sel.R
\name{model_sel.gwrm}
\alias{model_sel.gwrm}
\alias{model_sel}
\title{Generic method for model auto-selection.}
\usage{
\method{model_sel}{gwrm}(
  object,
  criterion = c("AIC"),
  threshold = 3,
  bw = NA,
  optim_bw = c("no", "AIC", "CV"),
  ...
)

model_sel(object, ...)
}
\arguments{
\item{object}{Geographically Weighted Model}

\item{criterion}{The model-selection method.
Currently there is only AIC available.}

\item{threshold}{The threshold of criterion changes. Default to 3.0.}

\item{bw}{The bandwidth used in selecting models.
If `NA` or missing, the bandwidth value will be derived
from the `gwrm` object.}

\item{optim_bw}{Whether optimize bandwidth after selecting models.
Avaliable values are `no`, `AIC`, and `CV`.
If `no` is specified, the bandwidth specified by argument `bw`
is used in calibrating selected models.}

\item{\dots}{Other arguments passing to implementation.}
}
\value{
A `gwrm` object.
}
\description{
Generic method for model auto-selection.
}
\section{Methods (by class)}{
\itemize{
\item \code{gwrm}: Model selection for basic GWR model
}}

\examples{
data(LondonHP)
model_sel(gwr_basic(PURCHASE~FLOORSZ+UNEMPLOY+PROF, LondonHP, 'AIC', TRUE),
          threshold = 100.0)

}
