% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_sel.R
\name{model_sel_view_circle}
\alias{model_sel_view_circle}
\title{Create circle view for model selection criterions}
\usage{
model_sel_view_circle(object)
}
\arguments{
\item{object}{An object of `modelselcritl` class}
}
\description{
Create circle view for model selection criterions
}
