% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwr_multiscale.R
\name{new_mgwr_config}
\alias{new_mgwr_config}
\title{Create an instance of MGWRConfig}
\usage{
new_mgwr_config(
  bw = NA_real_,
  adaptive = FALSE,
  kernel = c("gaussian", "exp", "bisquare", "tricube", "boxcar"),
  longlat = FALSE,
  p = 2,
  theta = 0,
  centered = FALSE,
  optim_bw = c("no", "AIC", "CV"),
  optim_threshold = 1e-05
)
}
\description{
Create an instance of MGWRConfig
}
