% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwr_basic.R
\name{plot.gwrm}
\alias{plot.gwrm}
\title{Plot the result of basic GWR model.}
\usage{
\method{plot}{gwrm}(x, y, ...)
}
\arguments{
\item{x}{A "gwrm" object.}

\item{y}{Column names to plot.
If it is missing or non-character value, all coefficient columns are plottd.}

\item{\dots}{Additional arguments passing to [sf::plot()].}
}
\description{
Plot the result of basic GWR model.
}
