% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_sel.R
\name{plot.modelselcritl}
\alias{plot.modelselcritl}
\alias{model_sel_view_circle}
\alias{model_sel_view_value}
\alias{model_sel_view_diff}
\title{Plot model selection critions (the circle view).}
\usage{
\method{plot}{modelselcritl}(object, view = c("circle", "value", "diff"), ymin, main, ...)

model_sel_view_circle(object, ...)

model_sel_view_value(object, ...)

model_sel_view_diff(object, ymin = -50, ...)
}
\arguments{
\item{object}{An object of `modelselcritl` class.}

\item{ymin}{The lower bound of y-axis.}

\item{main}{The main title.}

\item{\dots}{Additional parameters passing to [plot()].}
}
\description{
Plot model selection critions (the circle view).
}
\section{Functions}{
\itemize{
\item \code{model_sel_view_circle}: Create circle view for
model combinations in model selection.

\item \code{model_sel_view_value}: Create scatter plot for
model selection criterion values

\item \code{model_sel_view_diff}: Create scatter plot for
differences of model selection criterion values
}}

\examples{
data(LondonHP)
m <- gwr_basic(
  formula = PURCHASE ~ FLOORSZ + UNEMPLOY + PROF + BATH2 + BEDS2 +
      GARAGE1 + TYPEDETCH + TPSEMIDTCH + TYPETRRD + TYPEBNGLW +
      BLDPWW1 +BLDPOSTW + BLD60S + BLD70S + BLD80S + CENTHEAT,
  data = LondonHP,
  bw = "AIC",
  adaptive = TRUE
) |> model_sel(threshold = 10)
plot(m$model_sel)
plot(m$model_sel, view = "value")
plot(m$model_sel, view = "diff")

}
