% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwdr.R, R/gwr_basic.R, R/gwr_multiscale.R,
%   R/model_sel.R
\name{plot}
\alias{plot}
\alias{plot.gwdrm}
\alias{plot.gwrm}
\alias{plot.gwrmultiscalem}
\alias{plot.modelselcritl}
\title{Plot the result of basic GWR model.}
\usage{
\method{plot}{gwdrm}(x, y, ..., columns)

\method{plot}{gwdrm}(x, y, ..., columns)

\method{plot}{gwrm}(x, y, ..., columns)

\method{plot}{gwrmultiscalem}(x, y, ..., columns)

\method{plot}{modelselcritl}(x, y, ..., view = c("circle", "value", "diff"), ymin, main)
}
\arguments{
\item{x}{An object of `modelselcritl` class.}

\item{y}{Ignored.}

\item{\dots}{Additional parameters passing to [plot()].}

\item{columns}{Column names to plot.
If it is missing or non-character value, all coefficient columns are plottd.}

\item{view}{Name of view.}

\item{ymin}{The lower bound of y-axis.}

\item{main}{The main title.}
}
\description{
Plot the result of basic GWR model.

Plot the result of basic GWR model.

Plot the result of basic GWR model.

Plot the result of basic GWR model.

Plot model selection criterions (the circle view).
}
\examples{
data(LondonHP)
m <- gwr_basic(PURCHASE~FLOORSZ+UNEMPLOY, LondonHP, 64, TRUE)
plot(m)

data(LondonHP)
m <- gwr_basic(PURCHASE~FLOORSZ+UNEMPLOY, LondonHP, 64, TRUE)
plot(m)

data(LondonHP)
m <- gwr_basic(PURCHASE~FLOORSZ+UNEMPLOY, LondonHP, 64, TRUE)
plot(m)

data(LondonHP)
m <- gwr_multiscale(
 formula = PURCHASE ~ FLOORSZ + UNEMPLOY + PROF,
 data = LondonHP,
 config = rep(mgwr_config(36, TRUE, "bisquare", optim_bw = "AIC"), 4)
)
plot(m)

data(LondonHP)
m <- gwr_basic(
  PURCHASE ~ FLOORSZ + UNEMPLOY + PROF + BATH2 + BEDS2 +
      GARAGE1 + TYPEDETCH + TPSEMIDTCH + TYPETRRD + TYPEBNGLW +
      BLDPWW1 +BLDPOSTW + BLD60S + BLD70S + BLD80S + CENTHEAT,
  LondonHP, bw = "AIC", adaptive = TRUE
)
m <- model_sel(m, threshold = 10, bw = Inf, optim_bw = "AIC")
plot(m$model_sel)
plot(m$model_sel, view = "value")
plot(m$model_sel, view = "diff")

}
