% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwr_basic.R
\name{predict}
\alias{predict}
\alias{predict.gwrm}
\title{Predict on new locations.}
\usage{
\method{predict}{gwrm}(object, regression_points, ...)
}
\arguments{
\item{object}{A "gwrm" object.}

\item{regression_points}{Data of new locations.}

\item{\dots}{Additional arguments.}
}
\value{
A "gwrm" object.
}
\description{
Predict on new locations.
}
\examples{
data(LondonHP)
m <- gwr_basic(PURCHASE ~ FLOORSZ + UNEMPLOY + PROF, LondonHP, 64, TRUE)
predict(m, LondonHP)

}
