% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-GTDRConfig.R
\name{rep-GTDRConfig}
\alias{rep-GTDRConfig}
\alias{rep,GTDRConfig-method}
\alias{rep.int,GTDRConfig,numeric-method}
\title{Replicate GTDR config}
\usage{
\S4method{rep}{GTDRConfig}(x, ...)

\S4method{rep.int}{GTDRConfig,numeric}(x, times = 1)
}
\arguments{
\item{x}{A \linkS4class{GTDRConfig} object.}

\item{\dots}{Additional arguments.}

\item{times}{Replication times.}
}
\value{
A list of \linkS4class{GTDRConfig} objects.
}
\description{
Replicate GTDR config
}
\examples{
rep(gtdr_config(36, TRUE, "bisquare"), 4)

}
