% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-GWCorrConfig.R
\name{rep-GWCorrConfig}
\alias{rep-GWCorrConfig}
\alias{rep,GWCorrConfig-method}
\alias{rep.int,GWCorrConfig,numeric-method}
\title{Replicate MGWR config}
\usage{
\S4method{rep}{GWCorrConfig}(x, ...)

\S4method{rep.int}{GWCorrConfig,numeric}(x, times = 1)
}
\arguments{
\item{x}{A \linkS4class{GWCorrConfig} object.}

\item{\dots}{Additional arguments.}

\item{times}{Replication times.}
}
\value{
A list of \linkS4class{GWCorrConfig} objects.
}
\description{
Replicate MGWR config
}
\examples{
rep(gwcorr_config(36, TRUE, "bisquare"), 4)

}
