% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-GWDRConfig.R
\name{rep-GWDRConfig}
\alias{rep-GWDRConfig}
\alias{rep,GWDRConfig-method}
\alias{rep.int,GWDRConfig,numeric-method}
\title{Replicate MGWR config}
\usage{
\S4method{rep}{GWDRConfig}(x, ...)

\S4method{rep.int}{GWDRConfig,numeric}(x, times = 1)
}
\arguments{
\item{x}{A \linkS4class{GWDRConfig} object.}

\item{\dots}{Additional arguments.}

\item{times}{Replication times.}
}
\value{
A list of \linkS4class{GWDRConfig} objects.
}
\description{
Replicate MGWR config
}
\examples{
rep(gwdr_config(36, TRUE, "bisquare"), 4)

}
