% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwr_multiscale.R
\name{rep-MGWRConfig}
\alias{rep-MGWRConfig}
\alias{rep,MGWRConfig-method}
\alias{rep.int,MGWRConfig,numeric-method}
\title{Replicate MGWR config}
\usage{
\S4method{rep}{MGWRConfig}(x, ...)

\S4method{rep.int}{MGWRConfig,numeric}(x, times = 1)
}
\arguments{
\item{x}{A \linkS4class{MGWRConfig} object.}

\item{\dots}{Additional arguments.}

\item{times}{Replication times.}
}
\value{
A list of \linkS4class{MGWRConfig} objects.
}
\description{
Replicate MGWR config
}
