% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-GWDRConfig.R
\name{rep-SDRConfig}
\alias{rep-SDRConfig}
\alias{rep,SDRConfig-method}
\alias{rep.int,SDRConfig,numeric-method}
\title{Replicate SDR config}
\usage{
\S4method{rep}{SDRConfig}(x, ...)

\S4method{rep.int}{SDRConfig,numeric}(x, times = 1)
}
\arguments{
\item{x}{A \linkS4class{SDRConfig} object.}

\item{\dots}{Additional arguments.}

\item{times}{Replication times.}
}
\value{
A list of \linkS4class{SDRConfig} objects.
}
\description{
Replicate SDR config
}
\examples{
rep(sdr_config(36, TRUE, "bisquare"), 4)

}
