% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwr_basic.R
\name{residuals.gwrm}
\alias{residuals.gwrm}
\title{Get residuals of a basic GWR model.}
\usage{
\method{residuals}{gwrm}(object, ...)
}
\arguments{
\item{object}{A "gwrm" object.}

\item{\dots}{Additional arguments passing to [residuals()].}
}
\description{
Get residuals of a basic GWR model.
}
