% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwdr.R, R/gwr_basic.R, R/gwr_multiscale.R
\name{residuals}
\alias{residuals}
\alias{residuals.gwdrm}
\alias{residuals.gwrm}
\alias{residuals.gwrmultiscalem}
\title{Get residuals of a basic GWR model.}
\usage{
\method{residuals}{gwdrm}(object, ...)

\method{residuals}{gwrm}(object, ...)

\method{residuals}{gwrmultiscalem}(object, ...)
}
\arguments{
\item{object}{A "gwrmultiscalem" object.}

\item{\dots}{Additional arguments passing to [residuals()].}
}
\description{
Get residuals of a basic GWR model.

Get residuals of a basic GWR model.

Get residuals of a basic GWR model.
}
