% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdr.R
\name{sdr}
\alias{sdr}
\alias{plot.sdrm}
\alias{step.sdrm}
\alias{coef.sdrm}
\alias{fitted.sdrm}
\alias{residuals.sdrm}
\title{Geographically Weighted Density Regression (SDR)}
\usage{
sdr(
  formula,
  data,
  config = list(sdr_config()),
  optim_bw = c("no", "CV", "AIC"),
  optim_bw_threshold = 1e-06,
  optim_bw_step = 0.02,
  optim_bw_max_iter = 1e+06,
  parallel_method = c("no", "omp"),
  parallel_arg = c(0),
  verbose = FALSE
)

\method{plot}{sdrm}(x, y, ..., columns)

\method{step}{sdrm}(
  object,
  ...,
  criterion = c("AIC"),
  threshold = 3,
  config = list(sdr_config()),
  optim_bw = c("no", "CV", "AIC"),
  optim_bw_threshold = 1e-06,
  optim_bw_step = 0.02,
  optim_bw_max_iter = 1e+06
)

\method{coef}{sdrm}(object, ...)

\method{fitted}{sdrm}(object, ...)

\method{residuals}{sdrm}(object, ...)
}
\arguments{
\item{formula}{Regresison model.}

\item{data}{A \code{sf} objects.}

\item{config}{Dimension-specified weighting configuration.
If it is omitted, bandwidth settings will be inherited from \code{object}.
Otherwise, it must be a list of \code{SDRConfig} objects.
Its length can be 1 or equal to the number of columns of coordinates in \code{data}.
When its length is 1, elements will be duplicated for each column in coordinates.}

\item{optim_bw}{Whether optimize bandwidth after selecting models.
Avaliable values are \code{no}, \code{AIC}, and \code{CV}.
If \code{no} is specified, the bandwidth specified by argument \code{bw}
is used in calibrating selected models.}

\item{optim_bw_threshold}{Threshold of bandwidth optimization.}

\item{optim_bw_step}{Step size in bandwidth optimization.}

\item{optim_bw_max_iter}{Maximum of iteration in bandwidth optimization.}

\item{parallel_method}{Parallel method.}

\item{parallel_arg}{Parallel method argument.}

\item{verbose}{Whether to print additional information.
A higher value leads to more information.}

\item{x}{A "sdrm" object.}

\item{y}{Ignored.}

\item{\dots}{Additional arguments passing to \code{\link[=residuals]{residuals()}}.}

\item{columns}{Column names to plot.
If it is missing or non-character value, all coefficient columns are plottd.}

\item{object}{A "sdrm" object.}

\item{criterion}{The model-selection method.
Currently there is only AIC available.}

\item{threshold}{The threshold of criterion changes. Default to 3.0.}
}
\value{
A \code{sdrm} object.
}
\description{
Geographically Weighted Density Regression (SDR)
}
\section{Functions}{
\itemize{
\item \code{plot(sdrm)}: Plot the result of SDR model.

\item \code{step(sdrm)}: Model selection for SDR model

\item \code{coef(sdrm)}: Get coefficients of a SDR model.

\item \code{fitted(sdrm)}: Get fitted values of a SDR model.

\item \code{residuals(sdrm)}: Get residuals of a SDR model.

}}
\examples{
data(LondonHP)
sdr(PURCHASE ~ FLOORSZ + UNEMPLOY, LondonHP)

### Specific Bandwidth
sdr(PURCHASE ~ FLOORSZ + UNEMPLOY, LondonHP, list(
    sdr_config(0.2, TRUE, "gaussian"),
    sdr_config(0.2, TRUE, "gaussian")
))

### Optim Bandwidth
m <- sdr(PURCHASE ~ FLOORSZ + UNEMPLOY + PROF, LondonHP, list(
    sdr_config(0.618, TRUE, "gaussian")
), optim_bw = "CV", "AIC")
m

plot(m)

step(m, threshold = 100.0)

coef(m)

fitted(m)

residuals(m)

}
