% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-GWDRConfig.R
\name{sdr_config}
\alias{sdr_config}
\title{Create an instance of SDRConfig.}
\usage{
sdr_config(
  bw = 0.618,
  adaptive = TRUE,
  kernel = c("gaussian", "exp", "bisquare", "tricube", "boxcar")
)
}
\arguments{
\item{bw}{Bandwidth value.}

\item{adaptive}{Whether the bandwidth value is adaptive or not.}

\item{kernel}{Kernel function used.}
}
\description{
Create an instance of SDRConfig.
}
\examples{
sdr_config(36, TRUE, "bisquare")

}
