% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwdr.R, R/gwr_basic.R, R/model_sel.R
\name{step}
\alias{step}
\alias{step.gwdrm}
\alias{step.gwrm}
\alias{step.default}
\alias{step_view_circle}
\alias{step_view_value}
\alias{step_view_diff}
\title{Model selection for GWDR model}
\usage{
\method{step}{gwdrm}(
  object,
  criterion = c("AIC"),
  threshold = 3,
  config = list(gwdr_config()),
  optim_bw = c("no", "AIC", "CV"),
  optim_bw_threshold = 1e-06,
  optim_bw_step = 0.02,
  optim_bw_max_iter = 1e+06,
  ...
)

\method{step}{gwrm}(
  object,
  criterion = c("AIC"),
  threshold = 3,
  bw = NA,
  optim_bw = c("no", "AIC", "CV"),
  ...
)

step(object, ...)

\method{step}{default}(object, ...)

step_view_circle(object, ...)

step_view_value(object, ...)

step_view_diff(object, ymin = -50, ...)
}
\arguments{
\item{object}{Geographically Weighted Model}

\item{criterion}{The model-selection method.
Currently there is only AIC available.}

\item{threshold}{The threshold of criterion changes. Default to 3.0.}

\item{config}{Dimension-specified weighting configuration.
If it is omitted, bandwidth settings will be inherited from `object`.
Otherwise, it must be a list of `GWDRConfig` objects.
Its length can be 1 or equal to the number of columns of coordinates in `data`.
When its length is 1, elements will be duplicated for each column in coordinates.}

\item{optim_bw}{Whether optimize bandwidth after selecting models.
Avaliable values are `no`, `AIC`, and `CV`.
If `no` is specified, the bandwidth specified by argument `bw`
is used in calibrating selected models.}

\item{optim_bw_threshold}{Threshold of bandwidth optimization.}

\item{optim_bw_step}{Step size in bandwidth optimization.}

\item{optim_bw_max_iter}{Maximum of iteration in bandwidth optimization.}

\item{\dots}{Other arguments passing to implementation.}

\item{bw}{The bandwidth used in selecting models.
If `NA` or missing, the bandwidth value will be derived
from the `gwrm` object.}

\item{ymin}{The lower boundary of y-axis.}
}
\value{
A `gwrm` object.
}
\description{
Model selection for GWDR model

Model selection for basic GWR model

Generic method for model auto-selection.
}
\section{Functions}{
\itemize{
\item \code{step_view_circle()}: Create circle view for
model combinations in model selection.

\item \code{step_view_value()}: Create scatter plot for
model selection criterion values

\item \code{step_view_diff()}: Create scatter plot for
differences of model selection criterion values

}}
\examples{
data(LondonHP)
step(gwdr(PURCHASE~FLOORSZ+UNEMPLOY, LondonHP))

data(LondonHP)
step(gwr_basic(PURCHASE~FLOORSZ+UNEMPLOY+PROF, LondonHP, 'AIC', TRUE),
          threshold = 100.0)

}
\keyword{internal}
