% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwr_basic.R, R/gwr_multiscale.R
\name{coef}
\alias{coef}
\alias{coef.gwrm}
\alias{coef.gwrmultiscalem}
\title{Get coefficients of a basic GWR model.}
\usage{
\method{coef}{gwrm}(object, ...)

\method{coef}{gwrmultiscalem}(object, ...)
}
\arguments{
\item{object}{A "gwrmultiscalem" object.}

\item{\dots}{Additional arguments passing to [coef()].}
}
\description{
Get coefficients of a basic GWR model.

Get coefficients of a multiscale GWR model.
}
