% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-GTDRConfig.R
\name{gtdr_config}
\alias{gtdr_config}
\title{Create an instance of GTDRConfig.}
\usage{
gtdr_config(
  bw = 0.618,
  adaptive = TRUE,
  kernel = c("gaussian", "exp", "bisquare", "tricube", "boxcar", "localperiodical")
)
}
\arguments{
\item{bw}{Bandwidth value.}

\item{adaptive}{Whether the bandwidth value is adaptive or not.}

\item{kernel}{Kernel function used.}
}
\description{
Create an instance of GTDRConfig.
}
\examples{
gtdr_config(36, TRUE, "bisquare")

}
