% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwdr.R
\name{gwdr}
\alias{gwdr}
\alias{plot.gwdrm}
\alias{coef.gwdrm}
\alias{fitted.gwdrm}
\alias{residuals.gwdrm}
\title{Geographically Weighted Density Regression (GWDR)}
\usage{
gwdr(
  formula,
  data,
  config = list(gwdr_config()),
  optim_bw = c("no", "AIC", "CV"),
  optim_bw_threshold = 1e-06,
  optim_bw_step = 0.02,
  optim_bw_max_iter = 1e+06,
  parallel_method = c("no", "omp"),
  parallel_arg = c(0),
  verbose = FALSE
)

\method{plot}{gwdrm}(x, y, ..., columns)

\method{coef}{gwdrm}(object, ...)

\method{fitted}{gwdrm}(object, ...)

\method{residuals}{gwdrm}(object, ...)
}
\arguments{
\item{formula}{Regresison model.}

\item{data}{A `sf` objects.}

\item{config}{Dimension-specified weighting configuration.
It must be a list of `GWDRConfig` objects.
Its length can be 1 or equal to the number of columns of coordinates in `data`.
When its length is 1, elements will be duplicated for each column in coordinates.}

\item{optim_bw}{Whether optimize bandwidth after selecting models.
Avaliable values are `no`, `AIC`, and `CV`.
If `no` is specified, the bandwidth specified by argument `bw`
is used in calibrating selected models.}

\item{optim_bw_threshold}{Threshold of bandwidth optimization.}

\item{optim_bw_step}{Step size in bandwidth optimization.}

\item{optim_bw_max_iter}{Maximum of iteration in bandwidth optimization.}

\item{parallel_method}{Parallel method.}

\item{parallel_arg}{Parallel method argument.}

\item{x}{A "gwdrm" object.}

\item{y}{Ignored.}

\item{\dots}{Additional arguments passing to [residuals()].}

\item{columns}{Column names to plot.
If it is missing or non-character value, all coefficient columns are plottd.}

\item{object}{A "gwdrm" object.}
}
\value{
A `gwdrm` object.
}
\description{
Geographically Weighted Density Regression (GWDR)
}
\section{Functions}{
\itemize{
\item \code{plot(gwdrm)}: Plot the result of GWDR model.

\item \code{coef(gwdrm)}: Get coefficients of a GWDR model.

\item \code{fitted(gwdrm)}: Get fitted values of a GWDR model.

\item \code{residuals(gwdrm)}: Get residuals of a GWDR model.

}}
\examples{
data(LondonHP)
m1 <- gwdr(PURCHASE~FLOORSZ+UNEMPLOY, LondonHP)
m1

### Specific Bandwidth
m2 <- gwdr(PURCHASE~FLOORSZ+UNEMPLOY, LondonHP, list(
    gwdr_config(0.2, TRUE, "gaussian"),
    gwdr_config(0.2, TRUE, "gaussian")
))
m2

### Optim Bandwidth
m3 <- gwdr(PURCHASE~FLOORSZ+UNEMPLOY, LondonHP, list(
    gwdr_config(0.618, TRUE, "gaussian")
), optim_bw = "AIC")
m3

data(LondonHP)
m <- gwr_basic(PURCHASE~FLOORSZ+UNEMPLOY, LondonHP, 64, TRUE)
plot(m)

}
