% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwdr.R
\name{gwdr}
\alias{gwdr}
\alias{plot.gwdrm}
\alias{step.gwdrm}
\alias{coef.gwdrm}
\alias{fitted.gwdrm}
\alias{residuals.gwdrm}
\title{Geographically Weighted Density Regression (GWDR)}
\usage{
gwdr(
  formula,
  data,
  config = list(gwdr_config()),
  optim_bw = c("no", "AIC", "CV"),
  optim_bw_threshold = 1e-06,
  optim_bw_step = 0.02,
  optim_bw_max_iter = 1e+06,
  parallel_method = c("no", "omp"),
  parallel_arg = c(0),
  verbose = FALSE
)

\method{plot}{gwdrm}(x, y, ..., columns)

\method{step}{gwdrm}(
  object,
  ...,
  criterion = c("AIC"),
  threshold = 3,
  config = list(gwdr_config()),
  optim_bw = c("no", "AIC", "CV"),
  optim_bw_threshold = 1e-06,
  optim_bw_step = 0.02,
  optim_bw_max_iter = 1e+06
)

\method{coef}{gwdrm}(object, ...)

\method{fitted}{gwdrm}(object, ...)

\method{residuals}{gwdrm}(object, ...)
}
\arguments{
\item{formula}{Regresison model.}

\item{data}{A \code{sf} objects.}

\item{config}{Dimension-specified weighting configuration.
If it is omitted, bandwidth settings will be inherited from \code{object}.
Otherwise, it must be a list of \code{GWDRConfig} objects.
Its length can be 1 or equal to the number of columns of coordinates in \code{data}.
When its length is 1, elements will be duplicated for each column in coordinates.}

\item{optim_bw}{Whether optimize bandwidth after selecting models.
Avaliable values are \code{no}, \code{AIC}, and \code{CV}.
If \code{no} is specified, the bandwidth specified by argument \code{bw}
is used in calibrating selected models.}

\item{optim_bw_threshold}{Threshold of bandwidth optimization.}

\item{optim_bw_step}{Step size in bandwidth optimization.}

\item{optim_bw_max_iter}{Maximum of iteration in bandwidth optimization.}

\item{parallel_method}{Parallel method.}

\item{parallel_arg}{Parallel method argument.}

\item{verbose}{Whether to print additional information.
A higher value leads to more information.}

\item{x}{A "gwdrm" object.}

\item{y}{Ignored.}

\item{\dots}{Additional arguments passing to \code{\link[=residuals]{residuals()}}.}

\item{columns}{Column names to plot.
If it is missing or non-character value, all coefficient columns are plottd.}

\item{object}{A "gwdrm" object.}

\item{criterion}{The model-selection method.
Currently there is only AIC available.}

\item{threshold}{The threshold of criterion changes. Default to 3.0.}
}
\value{
A \code{gwdrm} object.
}
\description{
Geographically Weighted Density Regression (GWDR)
}
\section{Functions}{
\itemize{
\item \code{plot(gwdrm)}: Plot the result of GWDR model.

\item \code{step(gwdrm)}: Model selection for GWDR model

\item \code{coef(gwdrm)}: Get coefficients of a GWDR model.

\item \code{fitted(gwdrm)}: Get fitted values of a GWDR model.

\item \code{residuals(gwdrm)}: Get residuals of a GWDR model.

}}
\examples{
data(LondonHP)
gwdr(PURCHASE~FLOORSZ+UNEMPLOY, LondonHP)

### Specific Bandwidth
gwdr(PURCHASE~FLOORSZ+UNEMPLOY, LondonHP, list(
    gwdr_config(0.2, TRUE, "gaussian"),
    gwdr_config(0.2, TRUE, "gaussian")
))

### Optim Bandwidth
m <- gwdr(PURCHASE~FLOORSZ+UNEMPLOY, LondonHP, list(
    gwdr_config(0.618, TRUE, "gaussian")
), optim_bw = "AIC")
m

plot(m)

step(m)

coef(m)

fitted(m)

residuals(m)

}
