% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-GWDRConfig.R
\docType{class}
\name{GWDRConfig-class}
\alias{GWDRConfig-class}
\alias{GWDRConfig}
\title{An S4 class to set GWDR configurations}
\description{
An S4 class to set GWDR configurations
}
\section{Slots}{

\describe{
\item{\code{bw}}{Bandwidth value.}

\item{\code{adaptive}}{Whether the bandwidth value is adaptive or not.}

\item{\code{kernel}}{Kernel function used.}
}}

