% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-GWDRConfig.R
\docType{class}
\name{GWDRConfig-class}
\alias{GWDRConfig-class}
\alias{GWDRConfig}
\alias{gwdr_config}
\title{An S4 class to set GWDR configurations}
\usage{
gwdr_config(
  bw = 0.618,
  adaptive = TRUE,
  kernel = c("gaussian", "exp", "bisquare", "tricube", "boxcar")
)
}
\arguments{
\item{bw}{Bandwidth value.}

\item{adaptive}{Whether the bandwidth value is adaptive or not.}

\item{kernel}{Kernel function used.}
}
\description{
An S4 class to set GWDR configurations

Create an instance of GWDRConfig
}
\section{Functions}{
\itemize{
\item \code{gwdr_config()}: 

}}
\section{Slots}{

\describe{
\item{\code{bw}}{Bandwidth value.}

\item{\code{adaptive}}{Whether the bandwidth value is adaptive or not.}

\item{\code{kernel}}{Kernel function used.}
}}

\examples{
gwdr_config(36, TRUE, "bisquare")

}
