\name{GWmodel3-package}
\alias{GWmodel3-package}
\alias{GWmodel3}
\docType{package}
\title{
\packageTitle{GWmodel3}
}
\description{
\packageDescription{GWmodel3}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{GWmodel3}
\packageIndices{GWmodel3}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{GWmodel3}

Maintainer: \packageMaintainer{GWmodel3}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
