% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwr_basic.R
\name{gwr_basic}
\alias{gwr_basic}
\alias{plot.gwrm}
\alias{coef.gwrm}
\alias{fitted.gwrm}
\alias{residuals.gwrm}
\alias{predict.gwrm}
\title{Calibrate a basic GWR model}
\usage{
gwr_basic(
  formula,
  data,
  bw = NA,
  adaptive = FALSE,
  kernel = c("gaussian", "exp", "bisquare", "tricube", "boxcar"),
  longlat = FALSE,
  p = 2,
  theta = 0,
  hatmatrix = TRUE,
  parallel_method = c("no", "omp"),
  parallel_arg = c(0),
  verbose = FALSE
)

\method{plot}{gwrm}(x, y, ..., columns)

\method{coef}{gwrm}(object, ...)

\method{fitted}{gwrm}(object, ...)

\method{residuals}{gwrm}(object, ...)

\method{predict}{gwrm}(object, regression_points, verbose = FALSE, ...)
}
\arguments{
\item{formula}{Regresison model.}

\item{data}{A `sf` objects.}

\item{bw}{Either a value to set the size of bandwidth,
or one of the following characters to set the criterion for
bandwidth auto-optimization process.
- `AIC`
- `CV`
Note that if `NA` or other non-numeric value is setted,
this parameter will be reset to `Inf`.}

\item{adaptive}{Whether the bandwidth value is adaptive or not.}

\item{kernel}{Kernel function used.}

\item{longlat}{Whether the coordinates}

\item{p}{Power of the Minkowski distance,
default to 2, i.e., Euclidean distance.}

\item{theta}{Angle in radian to roate the coordinate system, default to 0.}

\item{hatmatrix}{If TRUE, great circle will be caculated.}

\item{parallel_method}{Parallel method.}

\item{parallel_arg}{Parallel method argument.}

\item{verbose}{Whether to print additional message.}

\item{x}{A "gwrm" object.}

\item{y}{Ignored.}

\item{\dots}{Additional arguments.}

\item{columns}{Column names to plot.
If it is missing or non-character value, all coefficient columns are plottd.}

\item{object}{A "gwrm" object.}

\item{regression_points}{Data of new locations.}
}
\value{
A `gwrm` object.
}
\description{
Calibrate a basic GWR model
}
\section{Functions}{
\itemize{
\item \code{plot(gwrm)}: Plot the result of basic GWR model.

\item \code{coef(gwrm)}: Get coefficients of a basic GWR model.

\item \code{fitted(gwrm)}: Get fitted values of a basic GWR model.

\item \code{residuals(gwrm)}: Get residuals of a basic GWR model.

\item \code{predict(gwrm)}: Predict on new locations.

}}
\examples{
data(LondonHP)

# Basic usage
gwr_basic(PURCHASE~FLOORSZ+UNEMPLOY, LondonHP, 64, TRUE)

# Bandwidth Optimization
m1 <- gwr_basic(PURCHASE~FLOORSZ+UNEMPLOY, LondonHP, 'AIC', TRUE)
m1

plot(m1)

coef(m1)

fitted(m1)

residuals(m1)

predict(m1, LondonHP)

}
