% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwr_basic.R
\name{gwr_basic}
\alias{gwr_basic}
\title{Calibrate a basic GWR model}
\usage{
gwr_basic(
  formula,
  data,
  bw = NA,
  adaptive = FALSE,
  kernel = c("gaussian", "exp", "bisquare", "tricube", "boxcar"),
  longlat = FALSE,
  p = 2,
  theta = 0,
  hatmatrix = TRUE,
  parallel_method = c("no", "omp"),
  parallel_arg = c(0)
)
}
\arguments{
\item{formula}{Regresison model.}

\item{data}{A `sf` objects.}

\item{bw}{Either a value to set the size of bandwidth,
or one of the following characters to set the criterion for
bandwidth auto-optimization process.
- `AIC`
- `CV`
Note that if `NA` or other non-numeric value is setted,
this parameter will be reset to `Inf`.}

\item{adaptive}{Whether the bandwidth value is adaptive or not.}

\item{kernel}{Kernel function used.}

\item{longlat}{Whether the coordinates}

\item{p}{Power of the Minkowski distance,
default to 2, i.e., Euclidean distance.}

\item{theta}{Angle in radian to roate the coordinate system, default to 0.}

\item{hatmatrix}{If TRUE, great circle will be caculated}

\item{parallel_method}{Parallel method.}

\item{parallel_arg}{Parallel method argument.}
}
\value{
A `gwrm` object.
}
\description{
Calibrate a basic GWR model
}
\examples{
data(LondonHP)

# Basic usage
gwr_basic(PURCHASE~FLOORSZ+UNEMPLOY, LondonHP, 64, TRUE)

# Bandwidth Optimization
gwr_basic(PURCHASE~FLOORSZ+UNEMPLOY, LondonHP, 'AIC', TRUE)

}
