% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwr_basic.R
\name{gwr_basic}
\alias{gwr_basic}
\alias{step.gwrm}
\alias{plot.gwrm}
\alias{coef.gwrm}
\alias{fitted.gwrm}
\alias{residuals.gwrm}
\alias{predict.gwrm}
\title{Calibrate a basic GWR model}
\usage{
gwr_basic(
  formula,
  data,
  bw = NA,
  adaptive = FALSE,
  kernel = c("gaussian", "exp", "bisquare", "tricube", "boxcar"),
  longlat = FALSE,
  p = 2,
  theta = 0,
  optim_bw_range = c(0, Inf),
  hatmatrix = TRUE,
  parallel_method = c("no", "omp", "cuda"),
  parallel_arg = c(0),
  verbose = FALSE
)

\method{step}{gwrm}(
  object,
  ...,
  criterion = c("AIC"),
  threshold = 3,
  bw = NA,
  optim_bw = c("no", "AIC", "CV")
)

\method{plot}{gwrm}(x, y, ..., columns)

\method{coef}{gwrm}(object, ...)

\method{fitted}{gwrm}(object, ...)

\method{residuals}{gwrm}(object, ...)

\method{predict}{gwrm}(object, regression_points, verbose = FALSE, ...)
}
\arguments{
\item{formula}{Regresison model.}

\item{data}{A \code{sf} objects.}

\item{bw}{The bandwidth used in selecting models.
If \code{NA} or missing, the bandwidth value will be derived
from the \code{gwrm} object.}

\item{adaptive}{Whether the bandwidth value is adaptive or not.}

\item{kernel}{Kernel function used.}

\item{longlat}{Whether the coordinates}

\item{p}{Power of the Minkowski distance,
default to 2, i.e., Euclidean distance.}

\item{theta}{Angle in radian to roate the coordinate system, default to 0.}

\item{optim_bw_range}{Bounds on bandwidth optimization, a vector of two numeric elements.
Set to \code{NA_real_} to enable default values selected by the algorithm.}

\item{hatmatrix}{If TRUE, great circle will be caculated.}

\item{parallel_method}{Parallel method.}

\item{parallel_arg}{Parallel method argument.}

\item{verbose}{Whether to print additional message.}

\item{object}{A "gwrm" object.}

\item{\dots}{Additional arguments.}

\item{criterion}{The model-selection method.
Currently there is only AIC available.}

\item{threshold}{The threshold of criterion changes. Default to 3.0.}

\item{optim_bw}{Whether optimize bandwidth after selecting models.
Avaliable values are \code{no}, \code{AIC}, and \code{CV}.
If \code{no} is specified, the bandwidth specified by argument \code{bw}
is used in calibrating selected models.}

\item{x}{A "gwrm" object.}

\item{y}{Ignored.}

\item{columns}{Column names to plot.
If it is missing or non-character value, all coefficient columns are plottd.}

\item{regression_points}{Data of new locations.}
}
\value{
A \code{gwrm} object.

A \code{gwrm} object.
}
\description{
Calibrate a basic GWR model
}
\details{
\subsection{Parallelization}{

Two parallel methods are provided to speed up basic GWR algorithm:
\itemize{
\item Multithreading (\code{omp})
\item NVIDIA GPU Computing (\code{cuda})
}

See the vignettes about parallelization to learn more about this topic.
}
}
\section{Functions}{
\itemize{
\item \code{step(gwrm)}: Model selection for basic GWR model

\item \code{plot(gwrm)}: Plot the result of basic GWR model.

\item \code{coef(gwrm)}: Get coefficients of a basic GWR model.

\item \code{fitted(gwrm)}: Get fitted values of a basic GWR model.

\item \code{residuals(gwrm)}: Get residuals of a basic GWR model.

\item \code{predict(gwrm)}: Predict on new locations.

}}
\examples{
data(LondonHP)

# Basic usage
gwr_basic(PURCHASE ~ FLOORSZ + UNEMPLOY, LondonHP, 64, TRUE)

# Bandwidth Optimization
m <- gwr_basic(PURCHASE ~ FLOORSZ + UNEMPLOY + PROF, LondonHP, 'AIC', TRUE)
m

step(m, threshold = 100.0, bw = Inf, optim_bw = "AIC")

plot(m)

coef(m)

fitted(m)

residuals(m)

predict(m, LondonHP)

}
\seealso{
\code{browseVignettes("")}
}
