% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwr_multiscale.R
\name{gwr_multiscale}
\alias{gwr_multiscale}
\title{Multiscale GWR}
\usage{
gwr_multiscale(
  formula,
  data,
  config = list(mgwr_config()),
  criterion = c("dCVR", "CVR"),
  hatmatrix = T,
  retry_times = 5,
  parallel_method = c("no", "omp"),
  parallel_arg = c(0)
)
}
\description{
Multiscale GWR
}
\examples{
data(LondonHP)
m <- gwr_multiscale(
 formula = PURCHASE ~ FLOORSZ + UNEMPLOY + PROF,
 data = LondonHP
)

# Specify more configurations
m <- gwr_multiscale(
 formula = PURCHASE ~ FLOORSZ + UNEMPLOY + PROF,
 data = LondonHP,
 config = rep(mgwr_config(adaptive = TRUE, kernel = "bisquare"))
)

}
