% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwr_multiscale.R
\name{gwr_multiscale}
\alias{gwr_multiscale}
\alias{plot.gwrmultiscalem}
\alias{coef.gwrmultiscalem}
\alias{fitted.gwrmultiscalem}
\alias{residuals.gwrmultiscalem}
\title{Multiscale GWR}
\usage{
gwr_multiscale(
  formula,
  data,
  config = list(mgwr_config()),
  criterion = c("CVR", "dCVR"),
  hatmatrix = T,
  retry_times = 5,
  max_iterations = 2000,
  parallel_method = c("no", "omp"),
  parallel_arg = c(0),
  verbose = FALSE
)

\method{plot}{gwrmultiscalem}(x, y, ..., columns)

\method{coef}{gwrmultiscalem}(object, ...)

\method{fitted}{gwrmultiscalem}(object, ...)

\method{residuals}{gwrmultiscalem}(object, ...)
}
\arguments{
\item{formula}{Regresison model.}

\item{data}{A `sf` objects.}

\item{config}{Parameter-specified weighting configuration.
It must be a list of `MGWRConfig` objects.
Its length can be 1 or equal to the number of independent variables.
When its length is 1, elements will be duplicated for independent variable.}

\item{criterion}{Convergence criterion of back-fitting algorithm.}

\item{hatmatrix}{If TRUE, great circle will be caculated.}

\item{retry_times}{The number times of continually optimizing each 
parameter-specific bandwidth even though it meets the criterion of convergence,
for avoiding sub-optimal choice due to illusion of convergence.}

\item{max_iterations}{Maximum number of iterations in the back-fitting procedure.}

\item{parallel_method}{Parallel method.}

\item{parallel_arg}{Parallel method argument.}

\item{verbose}{Output information level.
Can be either `FALSE` or integer values.
Higher values will leads to more output information.}

\item{x}{A "gwrmultiscalem" object.}

\item{y}{Ignored.}

\item{\dots}{Additional arguments passing to [residuals()].}

\item{columns}{Column names to plot.
If it is missing or non-character value, all coefficient columns are plottd.}

\item{object}{A "gwrmultiscalem" object.}
}
\value{
A `gwrmultiscalem` object.
}
\description{
Multiscale GWR
}
\section{Functions}{
\itemize{
\item \code{plot(gwrmultiscalem)}: Plot the result of basic GWR model.

\item \code{coef(gwrmultiscalem)}: Get coefficients of a multiscale GWR model.

\item \code{fitted(gwrmultiscalem)}: Get fitted values of a basic GWR model.

\item \code{residuals(gwrmultiscalem)}: Get residuals of a basic GWR model.

}}
\examples{
data(LondonHP)
m <- gwr_multiscale(
 formula = PURCHASE ~ FLOORSZ + UNEMPLOY + PROF,
 data = LondonHP
)

# Specify more configurations for all variables
m <- gwr_multiscale(
 formula = PURCHASE ~ FLOORSZ + UNEMPLOY + PROF,
 data = LondonHP,
 config = list(mgwr_config(adaptive = TRUE, kernel = "bisquare"))
)

# Specify more configurations for each variables
m <- gwr_multiscale(
 formula = PURCHASE ~ FLOORSZ + UNEMPLOY + PROF,
 data = LondonHP,
 config = list(
     mgwr_config(adaptive = TRUE, kernel = "bisquare"),
     mgwr_config(adaptive = TRUE, kernel = "bisquare"),
     mgwr_config(adaptive = TRUE, kernel = "bisquare"),
     mgwr_config(adaptive = TRUE, kernel = "bisquare")
 ))

data(LondonHP)
m <- gwr_multiscale(
 formula = PURCHASE ~ FLOORSZ + UNEMPLOY + PROF,
 data = LondonHP,
 config = rep(mgwr_config(36, TRUE, "bisquare", optim_bw = "AIC"), 4)
)
plot(m)

}
