% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwr_multiscale.R
\name{gwr_multiscale}
\alias{gwr_multiscale}
\title{Multiscale GWR}
\usage{
gwr_multiscale(
  formula,
  data,
  config = list(mgwr_config()),
  criterion = c("CVR", "dCVR"),
  hatmatrix = T,
  retry_times = 5,
  max_iterations = 2000,
  parallel_method = c("no", "omp"),
  parallel_arg = c(0),
  verbose = FALSE
)
}
\arguments{
\item{formula}{Regresison model.}

\item{data}{A `sf` objects.}

\item{config}{Parameter-specified weighting configuration.
It must be a list of `MGWRConfig` objects.
Its length can be 1 or equal to the number of independent variables.
When its length is 1, elements will be duplicated for independent variable.}

\item{criterion}{Convergence criterion of back-fitting algorithm.}

\item{hatmatrix}{If TRUE, great circle will be caculated.}

\item{retry_times}{The number times of continually optimizing each 
parameter-specific bandwidth even though it meets the criterion of convergence,
for avoiding sub-optimal choice due to illusion of convergence.}

\item{max_iterations}{Maximum number of iterations in the back-fitting procedure.}

\item{parallel_method}{Parallel method.}

\item{parallel_arg}{Parallel method argument.}

\item{verbose}{Output information level.
Can be either `FALSE` or integer values.
Higher values will leads to more output information.}
}
\description{
Multiscale GWR
}
\examples{
data(LondonHP)
m <- gwr_multiscale(
 formula = PURCHASE ~ FLOORSZ + UNEMPLOY + PROF,
 data = LondonHP
)

# Specify more configurations for all variables
m <- gwr_multiscale(
 formula = PURCHASE ~ FLOORSZ + UNEMPLOY + PROF,
 data = LondonHP,
 config = list(mgwr_config(adaptive = TRUE, kernel = "bisquare"))
)

# Specify more configurations for each variables
m <- gwr_multiscale(
 formula = PURCHASE ~ FLOORSZ + UNEMPLOY + PROF,
 data = LondonHP,
 config = list(
     mgwr_config(adaptive = TRUE, kernel = "bisquare"),
     mgwr_config(adaptive = TRUE, kernel = "bisquare"),
     mgwr_config(adaptive = TRUE, kernel = "bisquare"),
     mgwr_config(adaptive = TRUE, kernel = "bisquare")
 ))

}
