% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwss.R
\name{gwss}
\alias{gwss}
\alias{plot.gwssm}
\title{Geographically Weighted Summary Statistics}
\usage{
gwss(
  formula,
  data,
  bw,
  adaptive = FALSE,
  mode = c("average", "correlation"),
  quantile = FALSE,
  kernel = c("gaussian", "exp", "bisquare", "tricube", "boxcar"),
  longlat = FALSE,
  p = 2,
  theta = 0,
  parallel_method = c("no", "omp"),
  parallel_arg = c(0)
)

\method{plot}{gwssm}(x, y, ..., columns)
}
\arguments{
\item{formula}{Regresison model.}

\item{data}{A \code{sf} objects.}

\item{bw}{Bandwidth value}

\item{adaptive}{Whether the bandwidth value is adaptive or not.}

\item{mode}{Algorithm mode. Average or correlation.}

\item{quantile}{Whether to calculate local quantiles.}

\item{kernel}{Kernel function used.}

\item{longlat}{Whether the coordinates}

\item{p}{Power of the Minkowski distance,
default to 2, i.e., Euclidean distance.}

\item{theta}{Angle in radian to roate the coordinate system, default to 0.}

\item{parallel_method}{Parallel method.}

\item{parallel_arg}{Parallel method argument.}

\item{x}{A "gwssm" object}

\item{y}{Ignored}

\item{\dots}{Additional arguments passing to \code{\link[sf:plot]{sf::plot()}}.}

\item{columnsColumn}{names to plot.
If it is missing or non-character value, all coefficient columns are plotted.}
}
\value{
A \code{gwssm} object.
}
\description{
Geographically Weighted Summary Statistics
}
\section{Functions}{
\itemize{
\item \code{plot(gwssm)}: Plot the result of basic GWSS model.

}}
\examples{
data(LondonHP)
m <- gwss(~ PURCHASE + FLOORSZ + UNEMPLOY, LondonHP, 64, TRUE)
gwss(~ PURCHASE + FLOORSZ + UNEMPLOY, LondonHP, 64, TRUE, mode = "correlation")

plot(m)

}
