% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-MGWRConfig.R
\docType{class}
\name{MGWRConfig-class}
\alias{MGWRConfig-class}
\alias{MGWRConfig}
\title{An S4 class to set Multiscale GWR configurations}
\description{
An S4 class to set Multiscale GWR configurations
}
\section{Slots}{

\describe{
\item{\code{bw}}{Bandwidth value.}

\item{\code{adaptive}}{Whether the bandwidth value is adaptive or not.}

\item{\code{kernel}}{Kernel function used.}

\item{\code{longlat}}{Whether the coordinates.}

\item{\code{p}}{Power of the Minkowski distance,
default to 2, i.e., Euclidean distance.}

\item{\code{theta}}{Angle in radian to roate the coordinate system, default to 0.}

\item{\code{centered}}{A logical vector of length equalling to the number of predictors,
and note intercept is not included;
if the element is TRUE, the corresponding predictor will be centered.}

\item{\code{optim_bw}}{Whether optimize bandwidth after selecting models.
Avaliable values are \code{no}, \code{AIC}, and \code{CV}.
If \code{no} is specified, the bandwidth specified by argument \code{bw}
is used in calibrating selected models.}

\item{\code{optim_threshold}}{Threshold of bandwidth optimization.}
}}

