% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-MGWRConfig.R
\docType{class}
\name{MGWRConfig-class}
\alias{MGWRConfig-class}
\alias{MGWRConfig}
\alias{mgwr_config}
\title{An S4 class to set Multiscale GWR configurations}
\usage{
mgwr_config(
  bw = NA_real_,
  adaptive = FALSE,
  kernel = c("gaussian", "exp", "bisquare", "tricube", "boxcar"),
  longlat = FALSE,
  p = 2,
  theta = 0,
  centered = TRUE,
  optim_bw = c("AIC", "CV", "no"),
  optim_threshold = 1e-05
)
}
\description{
An S4 class to set Multiscale GWR configurations

Create an instance of MGWRConfig
}
\section{Functions}{
\itemize{
\item \code{mgwr_config}: 
}}

\section{Slots}{

\describe{
\item{\code{bw}}{Bandwidth Value}
}}

\examples{
mgwr_config(36, TRUE, "bisquare", optim_bw = "AIC")

}
