% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwdr.R
\name{model_sel.gwdrm}
\alias{model_sel.gwdrm}
\title{Model selection for GWDR model}
\usage{
\method{model_sel}{gwdrm}(
  object,
  criterion = c("AIC"),
  threshold = 3,
  config = list(gwdr_config()),
  optim_bw = c("no", "AIC", "CV"),
  optim_bw_threshold = 1e-06,
  optim_bw_step = 0.02,
  optim_bw_max_iter = 1e+06,
  ...
)
}
\description{
Model selection for GWDR model
}
