% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_sel.R
\name{model_sel}
\alias{model_sel}
\title{Generic method for model auto-selection.}
\usage{
model_sel(x, ...)
}
\arguments{
\item{x}{Geographically Weighted Model}

\item{\dots}{Other arguments passing to implementation.}
}
\description{
Generic method for model auto-selection.
}
