% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwr_basic.R, R/model_sel.R
\name{model_sel}
\alias{model_sel}
\alias{model_sel.gwrm}
\alias{model_sel_view_circle}
\alias{model_sel_view_value}
\alias{model_sel_view_diff}
\title{Model selection for basic GWR model}
\usage{
\method{model_sel}{gwrm}(
  object,
  criterion = c("AIC"),
  threshold = 3,
  bw = NA,
  optim_bw = c("no", "AIC", "CV"),
  ...
)

model_sel(object, ...)

model_sel_view_circle(object, ...)

model_sel_view_value(object, ...)

model_sel_view_diff(object, ymin = -50, ...)
}
\arguments{
\item{object}{Geographically Weighted Model}

\item{criterion}{The model-selection method.
Currently there is only AIC available.}

\item{threshold}{The threshold of criterion changes. Default to 3.0.}

\item{bw}{The bandwidth used in selecting models.
If `NA` or missing, the bandwidth value will be derived
from the `gwrm` object.}

\item{optim_bw}{Whether optimize bandwidth after selecting models.
Avaliable values are `no`, `AIC`, and `CV`.
If `no` is specified, the bandwidth specified by argument `bw`
is used in calibrating selected models.}

\item{\dots}{Other arguments passing to implementation.}
}
\value{
A `gwrm` object.
}
\description{
Model selection for basic GWR model

Generic method for model auto-selection.
}
\section{Functions}{
\itemize{
\item \code{model_sel_view_circle}: Create circle view for
model combinations in model selection.

\item \code{model_sel_view_value}: Create scatter plot for
model selection criterion values

\item \code{model_sel_view_diff}: Create scatter plot for
differences of model selection criterion values
}}

\examples{
data(LondonHP)
model_sel(gwr_basic(PURCHASE~FLOORSZ+UNEMPLOY+PROF, LondonHP, 'AIC', TRUE),
          threshold = 100.0)

}
