% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_sel.R
\name{step}
\alias{step}
\alias{step.default}
\alias{step_view_circle}
\alias{step_view_value}
\alias{step_view_diff}
\alias{plot.modelselcritl}
\title{Generic method for model auto-selection.}
\usage{
step(object, ...)

\method{step}{default}(object, ...)

step_view_circle(object, ...)

step_view_value(object, ...)

step_view_diff(object, ymin = -50, ...)

\method{plot}{modelselcritl}(x, y, ..., view = c("circle", "value", "diff"), ymin, main)
}
\arguments{
\item{object}{Geographically Weighted Model}

\item{\dots}{Additional parameters passing to \code{\link[=plot]{plot()}}.}

\item{ymin}{The lower bound of y-axis.}

\item{x}{An object of \code{modelselcritl} class.}

\item{y}{Ignored.}

\item{view}{Name of view.}

\item{main}{The main title.}
}
\description{
Generic method for model auto-selection.
}
\section{Functions}{
\itemize{
\item \code{step_view_circle()}: Create circle view for
model combinations in model selection.

\item \code{step_view_value()}: Create scatter plot for
model selection criterion values

\item \code{step_view_diff()}: Create scatter plot for
differences of model selection criterion values

\item \code{plot(modelselcritl)}: Plot model selection criterions.

}}
\examples{
data(LondonHP)
m <- gwr_basic(
  PURCHASE ~ FLOORSZ + UNEMPLOY + PROF + BATH2 + BEDS2 +
      GARAGE1 + TYPEDETCH + TPSEMIDTCH + TYPETRRD + TYPEBNGLW +
      BLDPWW1 +BLDPOSTW + BLD60S + BLD70S + BLD80S + CENTHEAT,
  LondonHP, bw = "AIC", adaptive = TRUE
)
m <- step(m, threshold = 10, bw = Inf, optim_bw = "AIC")
plot(m$step)
plot(m$step, view = "value")
plot(m$step, view = "diff")

}
\keyword{internal}
