% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwdr.R, R/gwr_basic.R, R/model_sel.R
\name{step}
\alias{step}
\alias{step.gwdrm}
\alias{step.gwrm}
\alias{step.default}
\alias{step_view_circle}
\alias{step_view_value}
\alias{step_view_diff}
\alias{plot.modelselcritl}
\title{Model selection for GWDR model}
\usage{
\method{step}{gwdrm}(
  object,
  criterion = c("AIC"),
  threshold = 3,
  config = list(gwdr_config()),
  optim_bw = c("no", "AIC", "CV"),
  optim_bw_threshold = 1e-06,
  optim_bw_step = 0.02,
  optim_bw_max_iter = 1e+06,
  ...
)

\method{step}{gwrm}(
  object,
  criterion = c("AIC"),
  threshold = 3,
  bw = NA,
  optim_bw = c("no", "AIC", "CV"),
  ...
)

step(object, ...)

\method{step}{default}(object, ...)

step_view_circle(object, ...)

step_view_value(object, ...)

step_view_diff(object, ymin = -50, ...)

\method{plot}{modelselcritl}(x, y, ..., view = c("circle", "value", "diff"), ymin, main)
}
\arguments{
\item{object}{Geographically Weighted Model}

\item{criterion}{The model-selection method.
Currently there is only AIC available.}

\item{threshold}{The threshold of criterion changes. Default to 3.0.}

\item{config}{Dimension-specified weighting configuration.
If it is omitted, bandwidth settings will be inherited from \code{object}.
Otherwise, it must be a list of \code{GWDRConfig} objects.
Its length can be 1 or equal to the number of columns of coordinates in \code{data}.
When its length is 1, elements will be duplicated for each column in coordinates.}

\item{optim_bw}{Whether optimize bandwidth after selecting models.
Avaliable values are \code{no}, \code{AIC}, and \code{CV}.
If \code{no} is specified, the bandwidth specified by argument \code{bw}
is used in calibrating selected models.}

\item{optim_bw_threshold}{Threshold of bandwidth optimization.}

\item{optim_bw_step}{Step size in bandwidth optimization.}

\item{optim_bw_max_iter}{Maximum of iteration in bandwidth optimization.}

\item{\dots}{Additional parameters passing to \code{\link[=plot]{plot()}}.}

\item{bw}{The bandwidth used in selecting models.
If \code{NA} or missing, the bandwidth value will be derived
from the \code{gwrm} object.}

\item{ymin}{The lower bound of y-axis.}

\item{x}{An object of \code{modelselcritl} class.}

\item{y}{Ignored.}

\item{view}{Name of view.}

\item{main}{The main title.}
}
\value{
A \code{gwrm} object.
}
\description{
Model selection for GWDR model

Model selection for basic GWR model

Generic method for model auto-selection.
}
\section{Functions}{
\itemize{
\item \code{step_view_circle()}: Create circle view for
model combinations in model selection.

\item \code{step_view_value()}: Create scatter plot for
model selection criterion values

\item \code{step_view_diff()}: Create scatter plot for
differences of model selection criterion values

\item \code{plot(modelselcritl)}: Plot model selection criterions.

}}
\examples{
data(LondonHP)
step(gwdr(PURCHASE~FLOORSZ+UNEMPLOY, LondonHP))

data(LondonHP)
step(gwr_basic(PURCHASE~FLOORSZ+UNEMPLOY+PROF, LondonHP, 'AIC', TRUE),
          threshold = 100.0)

data(LondonHP)
m <- gwr_basic(
  PURCHASE ~ FLOORSZ + UNEMPLOY + PROF + BATH2 + BEDS2 +
      GARAGE1 + TYPEDETCH + TPSEMIDTCH + TYPETRRD + TYPEBNGLW +
      BLDPWW1 +BLDPOSTW + BLD60S + BLD70S + BLD80S + CENTHEAT,
  LondonHP, bw = "AIC", adaptive = TRUE
)
m <- step(m, threshold = 10, bw = Inf, optim_bw = "AIC")
plot(m$step)
plot(m$step, view = "value")
plot(m$step, view = "diff")

}
\keyword{internal}
